refactor: update state and folder structure
This commit is a refactor of how the state is working leading to how the server state should work using rtk-query for that.

So it adds rtk-query to fetch the resume and keep the server state there. Also, add the bases to a new folder structure regarding to the `state` folder. Every selector and updater for  each resume section (`profile`, `contact`, `skills`, ...etc) is going to live inside `state/api/selectors-and-updaters`

I update the `Profile` section, from its type to its containers that consume the state, adding the flow to update the data into the server, so the other sections might follow the same pattern

Also, it changes the location for the unit types (`Size`, `Alignment`, `Weight`) to the `shared` folder

In the `server` folder, I create  `test-helpers/fuzzer-schemas` where I generate the fuzzers to test the `Resume` type splitter by its sections (`Profile`, `Contact`, `Skills`, ...etc) since these structure is going to grow a lot I think