Refactor UI input generation using string paths and adapters, and implement schema for elements (#15)
This is a refactor that creates a string path to generate inputs for UI elements such as title text, size, and color. An adapter renders the appropriate component based on the type of input required. The path can be used like an ID to achieve more granular control over the UI. The PR adds adapters for different types of inputs and a schema for each section. The PR also eliminates the use of the options.tsx and controls folder. However, more work is required to infer types from the string path.