Refactor Skills component and rename API function in slice
In this commit, we have refactored the Skills component to introduce SkillsComponent and utilized React.memo for memoization to optimize performance. Both the Chip and the TitleContainer components are memoized.

In the api slice, the useResumeContentUpdater function's internal variable was renamed from updateResume_ to update for clarity and better readability.