Migrate Storybook from version 6 to 7 and update Github Actions
This commit incorporates a substantial upgrade of Storybook from version 6 to 7. Modifications are made in the Storybook configuration which includes transitioning to the nextjs framework, updating addons, and turning off autodocs. The scripts in package.json for Storybook are also updated to reflect the changes necessary for the new version.

Another significant update is adding the /storybook-static directory to the .gitignore file which includes the static build artifacts of Storybook. Furthermore, README.md is updated to provide instructions for testing GitHub Actions locally.

Lastly, there are minor revisions and updates to various dependency versions in package.json. Collectively, these changes provide an efficient migration to Storybook 7 and align our project with the latest features and requirements. A new GitHub Action for deploying Storybook has also been added, facilitating continuous deployment and integration processes.